fix: some falcon-512 cleanup and fix fuzz tests
